[GEOS-10143] Factor out direct raster rendering in its own class. No simplification yet, just moving methods across classes
